+++ src/promptflow/promptflow/_sdk/schemas/_run.py
@@ -60,7 +60,7 @@ class RunSchema(YamlFileSchema):
     properties = fields.Dict(keys=fields.Str(), values=fields.Str(allow_none=True))
     # endregion: common fields
 
-    flow = LocalPathField(required=True)
+    flow = UnionField([LocalPathField(required=True), fields.Str(required=True)])
     # inputs field
     data = UnionField([LocalPathField(), RemotePathStr()])
     column_mapping = fields.Dict(keys=fields.Str)